+++ ExampleCodeBash/Services/HttpMethods/IHttpGet.sh
@@ -1,5 +1,8 @@
 cd $rootProjectDir/Services/HttpMethods
 
+## Dont overwrite
+test -f IHttpGet.cs && exit
+
 cat > IHttpGet.cs << EOM
 namespace "\$project_name".Services.HttpMethods
 {
@@ -8,8 +11,8 @@ namespace "\$project_name".Services.HttpMethods
     // IEnumerable<T> 
     "\$http_getAll_return_local" GetAll();
 
-    // T
-    "\$http_getById_return_local" GetById("\$http_getById_param_type_local" id"\$http_getById_param_local"); 
+    // T (int id)
+    "\$http_getById_return_local" GetById("\$http_getById_param_type_local" "\$http_getById_param_local"); 
   }
 }
 EOM
